examples(drill-downs): accessible dropdown menu
